+++ activerecord/lib/active_record/destroy_association_async_job.rb
@@ -4,23 +4,32 @@ module ActiveRecord
   class DestroyAssociationAsyncError < StandardError
   end
 
-  # Job to destroy the records associated with a destroyed record in background.
   class DestroyAssociationAsyncJob < ActiveJob::Base
     queue_as { ActiveRecord.queues[:destroy] }
 
     discard_on ActiveJob::DeserializationError
 
     def perform(
-      owner_model_name: nil, owner_id: nil,
-      association_class: nil, association_ids: nil, association_primary_key_column: nil,
-      ensuring_owner_was_method: nil
-    )
+          owner_model_name: nil,
+          owner_id: nil,
+          association_class: nil,
+          association_ids: nil,
+          association_primary_key_column: nil,
+          ensuring_owner_was_method: nil,
+          ensuring_owner_destroyed: true
+        )
       association_model = association_class.constantize
-      owner_class = owner_model_name.constantize
-      owner = owner_class.find_by(owner_class.primary_key.to_sym => owner_id)
 
-      if !owner_destroyed?(owner, ensuring_owner_was_method)
-        raise DestroyAssociationAsyncError, "owner record not destroyed"
+      # Handle the case when the `has_many` association is replaced with a new
+      # set. In such situation, respect the `dependent: :destroy_async`, and
+      # delete removed records in the background.
+      if ensuring_owner_destroyed
+        owner_class = owner_model_name.constantize
+        owner = owner_class.find_by(owner_class.primary_key.to_sym => owner_id)
+
+        if !owner_destroyed?(owner, ensuring_owner_was_method)
+          raise DestroyAssociationAsyncError, "owner record not destroyed"
+        end
       end
 
       association_model.where(association_primary_key_column => association_ids).find_each do |r|